+++ src/main/java/io/supertokens/pluginInterface/dashboard/DashboardStorage.java
@@ -1,5 +1,6 @@
 package io.supertokens.pluginInterface.dashboard;
 
+import io.supertokens.pluginInterface.ConfigFieldInfo;
 import io.supertokens.pluginInterface.Storage;
 import io.supertokens.pluginInterface.dashboard.exceptions.DuplicateEmailException;
 import io.supertokens.pluginInterface.dashboard.exceptions.DuplicateUserIdException;
@@ -7,6 +8,8 @@ import io.supertokens.pluginInterface.dashboard.exceptions.UserIdNotFoundExcepti
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.exceptions.TenantOrAppNotFoundException;
+
+import java.util.List;
 
 public interface DashboardStorage extends Storage {
 
@@ -35,4 +38,6 @@ public interface DashboardStorage extends Storage {
 
     // this function removes based on expired time, so we can use this to globally remove from a particular db.
     void revokeExpiredSessions() throws StorageQueryException;
+
+    List<ConfigFieldInfo> getPluginConfigFieldsInfo();
 }