+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [6.2.0] - 2024-05-24
+
+- Adds new class `ConfigFieldInfo` that represents a core config field
+- Adds new method `getPluginConfigFieldsInfo` to fetch the plugin config as json in `DashboardStorage`
+- Updates `TenantConfig` to support `null` and empty array state for `firstFactors`
+- Update `ThirdPartyConfig` to support `null` and empty array state for `providers`
+
 ## [6.1.0] - 2024-04-17
 
 - Adds `addTenantIdentifier` to the storage interface.